feat(order-payment): updated payment logic #16

+++ src/Services/Payments/Services.Payments.Application/BusinessLogics/CreatePayment/CreatePaymentDto.cs
@@ -3,7 +3,7 @@ public class CreatePaymentDto
 {
     public int TotalMoneyAmount { get; set; }
     public string? Description { get; set; }
-    public int OrderCode { get; set; }
+    public long OrderCode { get; set; }
     public string ReturnUrl { get; set; }
     public string CancelUrl { get; set; }
 }
